Moved all endpoints into their own file
Due to the size of all API endpoints, it started to become almost
impossible to keep them all in a single file. As a result, each class
was moved into their own file inside the endpoints directory.

This change also changes the way paths worked for each class. Despite
being nice having a single unified place to store all the magic
variables, it was difficult to keep track that all variables match
correctly to each function. Since they are used only once when calling
the wrapper, it is better to be mentioned only once in their specific
function.